- Restructured messages. - Added all required messages interfaces used for GridConnection but for SendJob. - Refactored classes from inline functions to header-source separation. - Removed IPCSearchPolicy. - Added new tests according to changes.

+++ src/boulderbreaker/resources/PCPerformance.h
@@ -4,7 +4,7 @@ struct CPUPerformance
 {
     int CPUCores;
     int CPUFrequency;
-    float CPULoad;
+    int CPULoad; // 0 - 100
 };
 
 struct PCPerformance
